Issue #33 explained and #52 refactored
