+++ issue17_test.go
@@ -17,9 +17,9 @@ var (
 func TestIssue17MergeWithOverwrite(t *testing.T) {
 	var something map[string]interface{}
 	if err := json.Unmarshal([]byte(request), &something); err != nil {
-		t.Errorf("Error while Unmarshalling maprequest %s", err)
+		t.Errorf("Error while Unmarshalling maprequest: %s", err)
 	}
 	if err := MergeWithOverwrite(&something, maprequest); err != nil {
-		t.Errorf("Error while merging %s", err)
+		t.Errorf("Error while merging: %s", err)
 	}
 }
